+++ packages/next-common/components/summary/newProposalButton/submitProposalPopup.js
@@ -3,127 +3,12 @@ import { NewProposalInnerPopup } from "../newProposalPopup";
 import { usePageProps } from "next-common/context/page";
 import { NewPreimageButton, NewProposalFromPreimageButton } from "./common";
 import SignerPopupWrapper from "next-common/components/popupWithSigner/signerPopupWrapper";
-import { QuickStart } from "next-common/components/preImages/createPreimagePopup";
-import { NewTreasuryReferendumInnerPopup } from "../newProposalQuickStart/createTreasuryProposalPopup";
-import { NewUSDxTreasuryReferendumInnerPopup } from "../newProposalQuickStart/createUSDxTreasuryProposalPopup";
-import { NewRemarkReferendumInnerPopup } from "../newProposalQuickStart/createSystemRemarkProposalPopup";
-import { CancelReferendumInnerPopup } from "../newProposalQuickStart/cancelReferendumInnerPopup";
-import {
-  SpendLocalTreasuryButton,
-  SpendUSDxTreasuryButton,
-  SpendDotOnAssetHubButton,
-  NewRemarkButton,
-  CancelReferendumButton,
-  KillReferendumButton,
-} from "next-common/components/preImages/createPreimagePopup/templateButtons";
-import { KillReferendumInnerPopup } from "../newProposalQuickStart/killReferendumInnerPopup";
-import { SpendDotOnAssetHubReferendumInnerPopup } from "../newProposalQuickStart/spendDotOnAssetHubPopup";
 import ForwardPopupProvider, {
   useForwardPopupContext,
 } from "next-common/context/forwardPopup";
-import { useChainSettings } from "next-common/context/chain";
 import Popup from "next-common/components/popup/wrapper/Popup";
 import { NewPreimageInnerPopup } from "next-common/components/preImages/newPreimagePopup";
-
-function SpendLocalTreasury() {
-  const { setForwardPopup } = useForwardPopupContext();
-  return (
-    <SpendLocalTreasuryButton
-      onClick={() => setForwardPopup(<NewTreasuryReferendumInnerPopup />)}
-    />
-  );
-}
-
-function SpendUSDxTreasury() {
-  const { setForwardPopup } = useForwardPopupContext();
-  const {
-    treasuryProposalTracks,
-    newProposalQuickStart: { usdxTreasuryProposal } = {},
-  } = useChainSettings();
-
-  if (!treasuryProposalTracks || !usdxTreasuryProposal) {
-    return null;
-  }
-
-  return (
-    <SpendUSDxTreasuryButton
-      onClick={() => setForwardPopup(<NewUSDxTreasuryReferendumInnerPopup />)}
-    />
-  );
-}
-
-function SpendDotOnAssetHub() {
-  const { setForwardPopup } = useForwardPopupContext();
-  const {
-    treasuryProposalTracks,
-    newProposalQuickStart: { spendDotOnAssetHubProposal } = {},
-  } = useChainSettings();
-
-  if (!treasuryProposalTracks || !spendDotOnAssetHubProposal) {
-    return null;
-  }
-
-  return (
-    <SpendDotOnAssetHubButton
-      onClick={() =>
-        setForwardPopup(<SpendDotOnAssetHubReferendumInnerPopup />)
-      }
-    />
-  );
-}
-
-function NewRemark() {
-  const { setForwardPopup } = useForwardPopupContext();
-  return (
-    <NewRemarkButton
-      onClick={() => setForwardPopup(<NewRemarkReferendumInnerPopup />)}
-    />
-  );
-}
-
-function CancelReferendum() {
-  const { setForwardPopup } = useForwardPopupContext();
-  const { newProposalQuickStart: { cancelReferendum } = {} } =
-    useChainSettings();
-
-  if (!cancelReferendum) {
-    return null;
-  }
-
-  return (
-    <CancelReferendumButton
-      onClick={() => setForwardPopup(<CancelReferendumInnerPopup />)}
-    />
-  );
-}
-
-function KillReferendum() {
-  const { setForwardPopup } = useForwardPopupContext();
-  const { newProposalQuickStart: { killReferendum } = {} } = useChainSettings();
-
-  if (!killReferendum) {
-    return null;
-  }
-
-  return (
-    <KillReferendumButton
-      onClick={() => setForwardPopup(<KillReferendumInnerPopup />)}
-    />
-  );
-}
-
-function ReferendaQuickStart() {
-  return (
-    <QuickStart>
-      <SpendLocalTreasury />
-      <SpendUSDxTreasury />
-      <SpendDotOnAssetHub />
-      <NewRemark />
-      <CancelReferendum />
-      <KillReferendum />
-    </QuickStart>
-  );
-}
+import ReferendaProposalQuickStart from "./referendaProposalQuickStart";
 
 function NewPreimage() {
   const { period } = usePageProps();
@@ -170,7 +55,7 @@ export default function SubmitProposalPopup({ onClose }) {
             <NewPreimage />
             <NewProposalFromPreImage />
           </div>
-          <ReferendaQuickStart />
+          <ReferendaProposalQuickStart />
         </Popup>
       </ForwardPopupProvider>
     </SignerPopupWrapper>
